requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-community==0.1.0
sentence-transformers==2.2.2
qdrant-client==1.7.0
pypdf==3.17.0
python-multipart==0.0.6
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
ollama==0.1.6
tiktoken==0.5.2
prometheus-client==0.19.0
redis==5.0.1